chore(app): moving StatsSection to separate file
